requirements.txt
torch
torchvision
scikit-learn
matplotlib
pillow
fastapi
uvicorn
